requirements/requirements.txt
requests
pyjwt
click
boto3
tabulate
jinja2
pyyaml